requirements.txt
scipy==1.4.1
matplotlib==3.3.0
numpy==1.18.5
py-ecg-detectors==1.0.2
pandas==1.1.0
scikit-learn==0.24.2
imblearn==0.0
PyWavelets==1.1.1
pickle-mixin==1.0.2
Counter==1.0.0
tqdm==4.61.1
opencv-python==4.5.2.54
